+++ src/gameData/cards/cardReroll.ts
@@ -1,7 +1,7 @@
 import { EventDice } from "../../events/eventDice";
 import { EventMessage } from "../../events/eventMessage";
-import { subroutine } from "../../routines/routineManager";
-import { askYesNo, execute, RoutineMove } from "../../routines/routines";
+import { Routine } from "../../routines/routine";
+import { RoutineMove } from "../../routines/routineMove";
 import { GameData } from "../gameData";
 import { Card } from "./card";
 
@@ -11,15 +11,15 @@ export class CardReroll extends Card{
         return yield* yield* new RoutineMove(data, dice);
     }
     private *roll(){
-        const dice = yield* execute(new EventDice(1));
-        yield* execute(new EventMessage(`${dice} が出ましたが振り直しますか？`));
-        const yes = yield* askYesNo();
+        const dice = yield* Routine.execute(new EventDice(1));
+        yield* Routine.execute(new EventMessage(`${dice} が出ましたが振り直しますか？`));
+        const yes = yield* Routine.askYesNo();
         yield 'end';
         yield 'end';
         return yes? yield* this.reroll(): dice;
     }
     private *reroll(){
-        const dice = yield* execute(new EventDice(1));
+        const dice = yield* Routine.execute(new EventDice(1));
         yield 'end';
         return dice;
     }
