+++ src/gameData/stations/station.ts
@@ -1,4 +1,4 @@
-import { subroutine } from "../../routines/routineManager";
+import { subroutine } from "../../routines/routine";
 import { Direction } from "../../utils/direction";
 import { Exportable } from "../../utils/exportable";
 import { SceneManager } from "../../utils/sceneManager";
