+++ src/gameData/stations/stationCard.ts
@@ -5,7 +5,7 @@ import { GameData } from "../gameData";
 import { Station, stationData } from "./station";
 import { cards } from "../cards/cards";
 import { Card } from "../cards/card";
-import { execute } from "../../routines/routines";
+import { Routine } from "../../routines/routine";
 
 export class StationCard extends Station{
     constructor(data : stationData, x: number = 0, y: number = 0, z: number = 0, id: number = -1){
@@ -13,7 +13,7 @@ export class StationCard extends Station{
     }
     *routine(gameData: GameData){
         yield new EventMessage('カード駅に止まった');
-        const id = yield* execute(new EventRoulette(StationCard.getCard, id => Card.data.get(id).name));
+        const id = yield* Routine.execute(new EventRoulette(StationCard.getCard, id => Card.data.get(id).name));
         yield 'end';
         yield 'end';
         const card = Card.get(id);
