+++ src/gameData/stations/stationPlus.ts
@@ -1,6 +1,6 @@
 import { EventMessage } from "../../events/eventMessage";
 import { EventRoulette } from "../../events/eventRoulette";
-import { execute } from "../../routines/routines";
+import { Routine } from "../../routines/routine";
 import { Util } from "../../utils/util";
 import { GameData } from "../gameData";
 import { Station, stationData } from "./station";
@@ -11,7 +11,7 @@ export class StationPlus extends Station{
     }
     *routine(gameData: GameData){
         yield new EventMessage('プラス駅に止まった');
-        const gain = yield* execute(new EventRoulette(StationPlus.gain(gameData), GameData.moneyToText));
+        const gain = yield* Routine.execute(new EventRoulette(StationPlus.gain(gameData), GameData.moneyToText));
         yield 'end';
         yield 'end';
         gameData.turnPlayer.money += gain;
